+++ Annulus/source/Annulus/World.cpp
@@ -73,8 +73,8 @@ namespace Annulus
 			mCollisionDetector->GenerateContacts(mColliders);
 
 			// Resolve contacts
-			mContactResolver->SetVelocityIterations(mContacts.size() * 4);
-			mContactResolver->SetPositionIterations(mContacts.size() * 4);
+			mContactResolver->SetVelocityIterations(static_cast<std::uint32_t>(mContacts.size() * 4));
+			mContactResolver->SetPositionIterations(static_cast<std::uint32_t>(mContacts.size() * 4));
 			mContactResolver->ResolveContacts(mContacts, seconds);
 
 			// Reset the time since last update.
